yalenta/requirements.txt
Django==3.1.7
sorl_thumbnail==12.7.0
django-debug-toolbar==3.2
Pillow==8.2.0
djangorestframework==3.12.4
whitenoise==5.2.0
psycopg2==2.8.6
django-cors-headers==3.7.0
boto3==1.17.60
django-storages==1.11.1
djoser==2.1.0
django-oauth-toolkit==1.5.0